FGW_homepage/RegDjangoServer/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djongo
dnspython
PyJWT
pymongo
pytz
sqlparse
typing_extensions
tzdata
